src/pyproject.toml
[build-system]
requires = [
    "setuptools >=65.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "api"
dynamic = ["version"]
readme = "README.md"
license = {text = 'Apache 2.0'}
description = "Python Project"
requires-python = ">=3.10"
dependencies = [
    "fastapi >=0.111.0, <1.0",
    "sqlmodel >=0.0.18, <0.1.0",
    "pydantic >=2.0, <3",
    "sniffio >1.2.0, <2",
    "greenlet>=3.0, <4",
    "httptools>=0.6, <1.0",
    "orjson>=3.9.11, <4",
    "uvloop>=0.18, <1",
    "certifi>=2024.2.2,<2025",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
    "chromadb>=1.0.20",
    "sentence-transformers>=5.1.0",
    "baml-py>=0.205.0",
]
[[project.authors]]
name = "Ben Epstein"
email = "ben.epstein97@gmail.com"

[project.optional-dependencies]
dev = [
    "coverage >=7.4.0, <8",
    "mypy >=1.10, <2",
    "pytest >=8.2, <9",
    "pytest-cov >=5.0.0, <6",
    "ruff >=0.4.5, <1.0",
]

[project.urls]
Documentation = "https://github.com/ben-epstein/python-project"

[tool.setuptools.dynamic]
version = {attr = "api.__version__"}

[tool.ruff]
line-length = 120

[tool.ruff.lint]
# Include rules for: pyFlakes, pycodestyle (Err/Warn), pyDocstyle, Isort, Flake8-FIXme, PyLinT, NumPY, debugger
# See rules and codes: https://docs.astral.sh/ruff/rules/
select = ["F", "E", "W", "D", "I001", "FIX", "PL", "NPY", "T100"]
dummy-variable-rgx = "_$|dummy"
ignore = ["FIX002", "D203", "D213"]


[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
mypy_path = ["api", "tests"]
exclude = ["dist", "build", ".venv"]

[tool.coverage.run]
source = ["api", "tests"]
omit = ["*__init__.py"]
parallel = true

[tool.coverage.report]
exclude_lines = [
    'pragma: no cover',
    'raise NotImplementedError'
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
